move native v8vm test to test/
