+++ test/native/test_data/test_new.js
