+++ test/native/token_test.go
@@ -1,4 +1,4 @@
-package vm
+package native
 
 import (
 	"testing"
