+++ test/v8vm/test_data/bignumber1.js
