+++ test/v8vm/test_data/console1.js
