+++ test/v8vm/test_data/danger.js
