+++ test/v8vm/test_data/func.js
