+++ test/v8vm/test_data/loop.js
