+++ test/v8vm/test_data/storage1.js
